Fix all warnings in odin module

+++ src/odin/maneuversbuilder.cc
@@ -283,7 +283,7 @@ std::list<Maneuver> ManeuversBuilder::Produce() {
       switch (node->type()) {
         case TripLeg_Node_Type_kStreetIntersection: {
           std::vector<std::pair<std::string, bool>> name_list;
-          for (size_t z = 0; z < node->intersecting_edge_size(); ++z) {
+          for (int z = 0; z < node->intersecting_edge_size(); ++z) {
             auto intersecting_edge = node->GetIntersectingEdge(z);
             for (const auto& name : intersecting_edge->name()) {
               std::pair<std::string, bool> cur_street = {name.value(), name.is_route_number()};
@@ -3758,7 +3758,7 @@ bool ManeuversBuilder::RampLeadsToHighway(Maneuver& maneuver) const {
   // Verify that the specified maneuver is a ramp
   if (maneuver.ramp()) {
     // Loop over edges
-    for (uint32_t node_index = maneuver.end_node_index(); node_index < trip_path_->GetLastNodeIndex();
+    for (int node_index = static_cast<int>(maneuver.end_node_index()); node_index < trip_path_->GetLastNodeIndex();
          ++node_index) {
       auto curr_edge = trip_path_->GetCurrEdge(node_index);
       if (curr_edge && (curr_edge->IsRampUse() || curr_edge->IsTurnChannelUse() ||
@@ -3782,7 +3782,7 @@ void ManeuversBuilder::SetTraversableOutboundIntersectingEdgeFlags(std::list<Man
   // Process each maneuver for traversable outbound intersecting edges
   for (Maneuver& maneuver : maneuvers) {
     bool found_first_edge_to_process = false;
-    for (int node_index = maneuver.begin_node_index(); node_index < maneuver.end_node_index();
+    for (int node_index = static_cast<int>(maneuver.begin_node_index()); node_index < static_cast<int>(maneuver.end_node_index());
          ++node_index) {
       if (!found_first_edge_to_process) {
         auto curr_edge = trip_path_->GetCurrEdge(node_index);
